Move `IndexRemover` to its own .hpp and .cpp file. (#1206)
